Fixed #78: Prompt to clean current directory if source is specified

+++ m2install.sh
@@ -1083,7 +1083,7 @@ then
 else
     if [[ "${SOURCE}" ]]
     then
-        if askConfirmation "Current directory is not empty. Do you want to clean current Directory (y/N)"
+        if [ "$(ls -A)" ] && askConfirmation "Current directory is not empty. Do you want to clean current Directory (y/N)"
         then
             CMD="ls -A | xargs rm -rf"
             runCommand
